@@ -3,34 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "babel-plugin-styled-components": "^1.10.7",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.6.5",
+    "babel-plugin-styled-components": "^2.0.3",
+    "bootstrap": "^5.1.3",
+    "core-js": "^3.21.1",
     "disqus-react": "^1.0.5",
-    "gatsby": "^2.12.0",
-    "gatsby-image": "^2.0.29",
-    "gatsby-plugin-manifest": "^2.0.17",
-    "gatsby-plugin-netlify-cms": "^4.1.1",
-    "gatsby-plugin-offline": "^2.0.23",
-    "gatsby-plugin-react-helmet": "^3.0.6",
-    "gatsby-plugin-sass": "^2.1.23",
-    "gatsby-plugin-sharp": "^2.0.20",
-    "gatsby-plugin-styled-components": "^3.1.19",
-    "gatsby-remark-prismjs": "^3.3.34",
-    "gatsby-source-filesystem": "^2.0.20",
-    "gatsby-transformer-remark": "^2.2.5",
-    "gatsby-transformer-sharp": "^2.1.13",
-    "gh-pages": "^2.0.1",
+    "gatsby": "^4.7.2",
+    "gatsby-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^4.7.0",
+    "gatsby-plugin-netlify-cms": "^6.7.0",
+    "gatsby-plugin-offline": "^5.7.0",
+    "gatsby-plugin-react-helmet": "^5.7.0",
+    "gatsby-plugin-sass": "^5.7.0",
+    "gatsby-plugin-sharp": "^4.7.0",
+    "gatsby-plugin-styled-components": "^5.7.0",
+    "gatsby-remark-prismjs": "^6.7.0",
+    "gatsby-source-filesystem": "^4.7.0",
+    "gatsby-transformer-remark": "^5.7.0",
+    "gatsby-transformer-sharp": "^4.7.0",
+    "gh-pages": "^3.2.3",
     "netlify-cms": "^2.5.1",
-    "node-sass": "^4.13.0",
+    "netlify-cms-app": "^2.15.66",
     "prismjs": "^1.19.0",
     "prop-types": "^15.7.2",
-    "react": "^16.8.2",
+    "react": "^17.0.2",
     "react-anchor-link-smooth-scroll": "^1.0.11",
-    "react-bootstrap": "^1.0.0-beta.5",
-    "react-dom": "^16.8.2",
-    "react-helmet": "^5.2.0",
-    "react-scripts": "2.1.5",
+    "react-bootstrap": "^2.1.2",
+    "react-dom": "^17.0.2",
+    "react-helmet": "^6.1.0",
+    "react-scripts": "5.0.0",
     "react-sticky-header": "^0.2.0",
     "react-sweet-progress": "^1.1.2",
     "styled-components": "^5.0.1"
@@ -52,5 +52,8 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "sass": "^1.49.8"
+  }
 }
